adding condition on database existence while connection termination

+++ shell_modules/postgresql_restore.sh
@@ -106,20 +106,26 @@ postgresql_restore() {
         log "postgresql (${DB}) getting dump file process ended (in success) in $(($PG_DMP_FILE_ELAPSED_TIME/60)) min $(($PG_DMP_FILE_ELAPSED_TIME%60)) sec"
 
         ## Get ride of existing database connection
-        PG_TERMINATE_CONN_DB_START_TIME=$(date +%s)
-        log "postgresql connection termination process started on host : ${POSTGRES_HOST} for Database : ${DB}"
-        PG_TERMINATE_CONN_RESULT=$(psql -h ${POSTGRES_HOST} -U ${POSTGRES_USERNAME} -d ${DB} -c "SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = '${DB}' AND pid <> pg_backend_pid();")
-        
-        if [ $? -ne 0 ]; then
+
+        # Check if the database exists
+        if psql -h ${POSTGRES_HOST} -U ${POSTGRES_USERNAME} -d postgres -tAc "SELECT 1 FROM pg_database WHERE datname='$DB'" | grep -q 1; then
+            PG_TERMINATE_CONN_DB_START_TIME=$(date +%s)
+            log "postgresql connection termination process started on host : ${POSTGRES_HOST} for Database : ${DB}"
+            PG_TERMINATE_CONN_RESULT=$(psql -h ${POSTGRES_HOST} -U ${POSTGRES_USERNAME} -d ${DB} -c "SELECT pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = '${DB}' AND pid <> pg_backend_pid();")
+            
+            if [ $? -ne 0 ]; then
+                PG_TERMINATE_CONN_ELAPSED_TIME=$(( $(date +%s)-${PG_TERMINATE_CONN_DB_START_TIME} ))
+                error "postgresql (${DB}) connection termination process ended (in error) in $(($PG_TERMINATE_CONN_ELAPSED_TIME/60)) min $(($PG_TERMINATE_CONN_ELAPSED_TIME%60)) sec"
+                error $PG_TERMINATE_CONN_RESULT
+                PG_RESTORE_SUCCESS=false
+                return 1
+                break
+            fi
             PG_TERMINATE_CONN_ELAPSED_TIME=$(( $(date +%s)-${PG_TERMINATE_CONN_DB_START_TIME} ))
-            error "postgresql (${DB}) connection termination process ended (in error) in $(($PG_TERMINATE_CONN_ELAPSED_TIME/60)) min $(($PG_TERMINATE_CONN_ELAPSED_TIME%60)) sec"
-            error $PG_TERMINATE_CONN_RESULT
-            PG_RESTORE_SUCCESS=false
-            return 1
-            break
+            log "postgresql (${DB}) connection termination process ended (in success) in $(($PG_TERMINATE_CONN_ELAPSED_TIME/60)) min $(($PG_TERMINATE_CONN_ELAPSED_TIME%60)) sec"
         fi
-        PG_TERMINATE_CONN_ELAPSED_TIME=$(( $(date +%s)-${PG_TERMINATE_CONN_DB_START_TIME} ))
-        log "postgresql (${DB}) connection termination process ended (in success) in $(($PG_TERMINATE_CONN_ELAPSED_TIME/60)) min $(($PG_TERMINATE_CONN_ELAPSED_TIME%60)) sec"
+
+
 
         ## restore database from restic restore
         PG_RESTORE_DB_START_TIME=$(date +%s)
